refactor(redesign-plugin-menu): naming
