+++ packages/editor/src/core/components/add-plugin-modal/plugin-menu-item.tsx
